chore(flags): Remove first-event-severity-calculation feature and dead code
This feature was only rolled out to early adopters and never GA'd.
Since the `first-event-severity-calculation` flag was always checked
after `seer-based-priority`, and was disabled for non-EA orgs, the
entire severity score calculation code path was effectively dead for
the vast majority of users.

Removes:
- Feature flag registration and check
- `_get_severity_metadata_for_group` function
- `_get_severity_score` function and Seer API call
- `make_severity_score_request`, `SeverityScoreRequest`
- `severity_connection_pool`, `update_severity_error_count`
- Related options (seer-timeout, rate limits, circuit breaker, etc.)
- test_severity.py (entirely dead)
- Feature decorators from test_priority.py and test_occurrence_consumer.py

Keeps:
- `_get_priority_for_group` (still assigns priority based on log level)
- `HIGH_SEVERITY_THRESHOLD` (used by above)
- `issues.severity.skip-seer-requests` killswitch (referenced elsewhere)
